Control Packet Execution Logic
Improved the logic for updating attribute values when a Control Packet is received.
